tests fixed and new added
test_UoW.py: added

test_user_repository.py: added

UoW.py: bug fixed

models.py: changed type of id to Integer (cause - problems with autoincrement in sqlite)
changes in migrations because of changes in models.py

+++ alembic/versions/seeding_admin.py
@@ -13,7 +13,7 @@ depends_on: Union[str, Sequence[str], None] = None
 
 users_table = sa.table(
     "users",
-    sa.column("id", sa.BigInteger),
+    sa.column("id", sa.Integer),
     sa.column("username", sa.String),
     sa.column("email", sa.String),
     sa.column("hashed_password", sa.String),
